Calibrate
The cube is now in the middle and the calibrate scene is new. The new scene allows you to calibrate the position. For that: first select a video and then pause it in the decoder scene. Then click calibrate and position your device according to the transparent image in your background. When that is done, go back to the decoder scene and select the video and now the video is playing accordingly.